requirements.txt
discord.py[voice]>=1.7.3
dislash
